+++ src/Polyfills/TaskExtensions.cs
@@ -5,7 +5,7 @@ namespace RunScript;
 
 using System.Diagnostics;
 
-internal static class Polyfills
+internal static class TaskExtensions
 {
 #if !NET5_0_OR_GREATER
     public static async Task WaitForExitAsync(this Process process, CancellationToken cancellationToken)